+++ deals-spring-boot-autoconfigure/pom.xml
@@ -28,6 +28,12 @@
             <artifactId>deals</artifactId>
             <optional>true</optional>
         </dependency>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+			<scope>test</scope>
+		</dependency>
     </dependencies>
 
 </project>
