--- sample-deals-app/src/test/java/io/agilehandy/sample/SampleApplicationTests.java
@@ -1,16 +1,0 @@
-package io.agilehandy.sample;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SampleApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
